# docs/swagger.yaml
definitions:
  R.ResponseEntity:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  dto.CpuInfo:
    properties:
      logicalCount:
        type: integer
      modelName:
        type: string
      percentage:
        type: number
      physicalCount:
        type: integer
    type: object
  dto.DiskInfo:
    properties:
      available:
        type: string
      percentage:
        type: number
      total:
        type: string
      used:
        type: string
    type: object
  dto.HostInfo:
    properties:
      bootTime:
        type: string
      dockerVersion:
        type: string
      kernel:
        type: string
      platForm:
        type: string
    type: object
  dto.MemoryInfo:
    properties:
      available:
        type: string
      percentage:
        type: number
      total:
        type: string
      used:
        type: string
    type: object
  dto.MicroServiceSytemInfo:
    properties:
      gateways:
        type: integer
      network:
        type: string
      registry:
        type: string
      serviceGroups:
        type: integer
      serviceInstances:
        $ref: '#/definitions/dto.ServiceInstancesInfo'
    type: object
  dto.ServiceInstancesInfo:
    properties:
      healthy:
        type: integer
      total:
        type: integer
      unHealthy:
        type: integer
    type: object
  dto.SystemPerfDTO:
    properties:
      cpu:
        $ref: '#/definitions/dto.CpuInfo'
      disk:
        $ref: '#/definitions/dto.DiskInfo'
      host:
        $ref: '#/definitions/dto.HostInfo'
      memory:
        $ref: '#/definitions/dto.MemoryInfo'
      mssystem:
        $ref: '#/definitions/dto.MicroServiceSytemInfo'
    type: object
  dto.UserInfoDTO:
    properties:
      avatar:
        type: string
      permissions:
        items:
          type: string
        type: array
      username:
        type: string
    type: object
  entity.UserEntity:
    properties:
      id:
        type: string
      password:
        type: string
      role:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  gin.H:
    additionalProperties: true
    type: object
  models.Traffic:
    properties:
      day:
        type: integer
      hour:
        type: integer
      level:
        type: string
      minute:
        type: integer
      month:
        type: integer
      number:
        type: integer
      year:
        type: integer
    type: object
...
info:
  contact:
    email: zengtao0618@163.com
    name: 曾涛
  description: 监控微服务系统中各个微服务实例的运行时状态以及所在容器的性能参数
  termsOfService: http://swagger.io/terms/
  title: FFR4MS Docker监控模块Fr-Docker
  version: "1.0"
paths:
...
  /container/logs:
    get:
      description: 获取微服务容器运行日志
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 容器的IP地址
        in: query
        name: ip
        required: true
        type: string
      - description: 日志行数，默认100
        in: query
        name: tail
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回微服务容器运行日志
          schema:
            allOf:
            - $ref: '#/definitions/R.ResponseEntity'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: 返回失败信息
          schema:
            $ref: '#/definitions/R.ResponseEntity'
      security:
      - ApiKeyAuth: []
      summary: 获取微服务容器运行日志
      tags:
      - 微服务容器操作
  /container/monitorLogs:
    get:
      description: 获取微服务容器监控日志
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 容器的IP地址
        in: query
        name: ip
        required: true
        type: string
      - description: 日志行数，默认100
        in: query
        name: tail
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回微服务容器监控日志
          schema:
            allOf:
            - $ref: '#/definitions/R.ResponseEntity'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: 返回失败信息
          schema:
            $ref: '#/definitions/R.ResponseEntity'
      security:
      - ApiKeyAuth: []
      summary: 获取微服务容器监控日志
      tags:
      - 微服务容器操作
  /container/traffic:
    get:
      description: 获取微服务容器访问流量
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 容器的IP地址
        in: query
        name: ip
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回微服务容器访问流量
          schema:
            allOf:
            - $ref: '#/definitions/R.ResponseEntity'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Traffic'
                  type: array
              type: object
        "400":
          description: 返回失败信息
          schema:
            $ref: '#/definitions/R.ResponseEntity'
      security:
      - ApiKeyAuth: []
      summary: 获取微服务容器访问流量
      tags:
      - 微服务容器操作
  /perf/container:
    get:
      description: 获取微服务容器性能参数
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回微服务容器性能参数
          schema:
            allOf:
            - $ref: '#/definitions/R.ResponseEntity'
            - properties:
                data:
                  $ref: '#/definitions/dto.SystemPerfDTO'
              type: object
        "400":
          description: 返回失败信息
          schema:
            $ref: '#/definitions/R.ResponseEntity'
      security:
      - ApiKeyAuth: []
      summary: 获取微服务容器性能参数
      tags:
      - 系统性能操作
  /perf/system:
    get:
      description: 获取服务器性能参数
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 容器的IP地址
        in: query
        name: ip
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回服务器性能参数
          schema:
            allOf:
            - $ref: '#/definitions/R.ResponseEntity'
            - properties:
                data:
                  $ref: '#/definitions/dto.SystemPerfDTO'
              type: object
        "400":
          description: 返回失败信息
          schema:
            $ref: '#/definitions/R.ResponseEntity'
      security:
      - ApiKeyAuth: []
      summary: 获取服务器性能参数
      tags:
      - 系统性能操作
  /user/info:
    get:
      description: 获取用户信息
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回用户信息
          schema:
            allOf:
            - $ref: '#/definitions/R.ResponseEntity'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserInfoDTO'
              type: object
        "400":
          description: 返回失败信息
          schema:
            $ref: '#/definitions/R.ResponseEntity'
      security:
      - ApiKeyAuth: []
      summary: 获取用户信息
      tags:
      - 用户操作
  /user/login:
    post:
      consumes:
      - application/json
      description: 用户登录
      parameters:
      - description: 登录用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/entity.UserEntity'
      produces:
      - application/json
      responses:
        "200":
          description: 返回登录Token
          schema:
            allOf:
            - $ref: '#/definitions/R.ResponseEntity'
            - properties:
                data:
                  $ref: '#/definitions/gin.H'
              type: object
        "400":
          description: 返回失败信息
          schema:
            $ref: '#/definitions/R.ResponseEntity'
      security:
      - ApiKeyAuth: []
      summary: 用户登录
      tags:
      - 用户操作
  /user/logout:
    post:
      description: 退出登录
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回登录Token
          schema:
            allOf:
            - $ref: '#/definitions/R.ResponseEntity'
            - properties:
                data:
                  $ref: '#/definitions/gin.H'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 退出登录
      tags:
      - 用户操作
  /user/register:
    post:
      consumes:
      - application/json
      description: 用户注册
      parameters:
      - description: 注册用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/entity.UserEntity'
      produces:
      - application/json
      responses:
        "200":
          description: 返回成功信息
          schema:
            $ref: '#/definitions/R.ResponseEntity'
        "400":
          description: 返回失败信息
          schema:
            $ref: '#/definitions/R.ResponseEntity'
      security:
      - ApiKeyAuth: []
      summary: 用户注册
      tags:
      - 用户操作
swagger: "2.0"